consolidate in es5/index.js
