+++ lib/connect.js
@@ -8,7 +8,7 @@ const EncodingParametersImpl = require('./encodingparameters');
 const ConstantIceServerSource = require('./iceserversource/constant');
 const NTSIceServerSource = require('./iceserversource/nts');
 const LocalParticipant = require('./localparticipant');
-const { LocalAudioTrack, LocalDataTrack, LocalVideoTrack } = require('./localtracks');
+const { LocalAudioTrack, LocalDataTrack, LocalVideoTrack } = require('./media/track/es5');
 const NetworkQualityConfigurationImpl = require('./networkqualityconfiguration');
 const Room = require('./room');
 const SignalingV2 = require('./signaling/v2');
