+++ lib/index.js
@@ -31,15 +31,15 @@ Object.defineProperties(Video, {
   },
   LocalAudioTrack: {
     enumerable: true,
-    value: require('./localtracks').LocalAudioTrack
+    value: require('./media/track/es5').LocalAudioTrack
   },
   LocalDataTrack: {
     enumerable: true,
-    value: require('./localtracks').LocalDataTrack
+    value: require('./media/track/es5').LocalDataTrack
   },
   LocalVideoTrack: {
     enumerable: true,
-    value: require('./localtracks').LocalVideoTrack
+    value: require('./media/track/es5').LocalVideoTrack
   },
   version: {
     enumerable: true,
