+++ lib/localparticipant.js
@@ -4,7 +4,7 @@ const { MediaStreamTrack } = require('@twilio/webrtc');
 const { asLocalTrack, asLocalTrackPublication, trackClass } = require('./util');
 const { typeErrors: E, trackPriority } = require('./util/constants');
 const { validateLocalTrack } = require('./util/validate');
-const { LocalAudioTrack, LocalDataTrack, LocalVideoTrack } = require('./localtracks');
+const { LocalAudioTrack, LocalDataTrack, LocalVideoTrack } = require('./media/track/es5');
 const LocalAudioTrackPublication = require('./media/track/localaudiotrackpublication');
 const LocalDataTrackPublication = require('./media/track/localdatatrackpublication');
 const LocalVideoTrackPublication = require('./media/track/localvideotrackpublication');
