--- lib/localtrackses6.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = {
-  LocalAudioTrack: require('./media/track/localaudiotrack'),
-  LocalVideoTrack: require('./media/track/localvideotrack'),
-  LocalDataTrack: require('./media/track/localdatatrack'),
-};
