+++ lib/media/track/es5/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  LocalAudioTrack: require('./localaudiotrack'),
+  LocalVideoTrack: require('./localvideotrack'),
+  LocalDataTrack: require('./localdatatrack'),
+};
